Create OpenToFirstBlank
This script loads on Open of Google Sheet and will search column A for the first blank cell (="") and make that the active cell. This shifts the window for the user down to that cell. This is good for lists that build over time. I add it to a lot of sheets and freeze the top row(s). Makes things much more user friendly.